Change in handling compiling inside vim.
Rather than using vim output window, now a new split will be opened in
which the compiling process will run. The benefit of this solution is
colored highlighting in the running compiler and cleaner code for the
compiler script.

Currently probably not as stable as the previous version. I will
continue to work on it.